Fix all review issues: types, parsing, output format, scripts
- Harmonize integer widths: u32 for node data, i64 for dist (all langs)
- Standardize output: read=X.Xms compute=X.Xms checksum=N visited=N
- C++: uint32_t throughout, int64_t for dist
- Rust: byte-level parsing (skip UTF-8 validation)
- Haskell: Word32 for node arrays, showFFloat for output
- Lean: ByteArray parsing (not String), fixed dead code, fmtMs helper
- Drop dense+large graph (689MB), cap at dense+100K
- Add build.sh and bench.sh
- Remove committed baseline binaries
- Update .gitignore for baseline and new patterns

All 4 languages produce matching checksums across sparse/medium/dense.

+++ cpp/bfs_barebones.cpp
@@ -22,30 +22,30 @@ int main(int argc, char **argv) {
     FILE *f = fopen(argv[1], "r");
     if (!f) { perror("fopen"); return 1; }
 
-    int n, m;
-    fscanf(f, "%d %d", &n, &m);
+    uint32_t n, m;
+    fscanf(f, "%u %u", &n, &m);
 
     // CSR-style adjacency: offsets + flat neighbor array
     // First pass: count degrees
-    int *deg = (int *)calloc(n, sizeof(int));
-    int *eu = (int *)malloc(m * sizeof(int));
-    int *ev = (int *)malloc(m * sizeof(int));
-    for (int i = 0; i < m; i++) {
-        fscanf(f, "%d %d", &eu[i], &ev[i]);
+    uint32_t *deg = (uint32_t *)calloc(n, sizeof(uint32_t));
+    uint32_t *eu = (uint32_t *)malloc(m * sizeof(uint32_t));
+    uint32_t *ev = (uint32_t *)malloc(m * sizeof(uint32_t));
+    for (uint32_t i = 0; i < m; i++) {
+        fscanf(f, "%u %u", &eu[i], &ev[i]);
         deg[eu[i]]++;
         deg[ev[i]]++;
     }
     fclose(f);
 
     // Build adjacency arrays
-    int *offset = (int *)malloc((n + 1) * sizeof(int));
+    uint32_t *offset = (uint32_t *)malloc((n + 1) * sizeof(uint32_t));
     offset[0] = 0;
-    for (int i = 0; i < n; i++) offset[i + 1] = offset[i] + deg[i];
-    int total_edges = offset[n];
-    int *adj = (int *)malloc(total_edges * sizeof(int));
-    int *pos = (int *)calloc(n, sizeof(int));
-    for (int i = 0; i < m; i++) {
-        int u = eu[i], v = ev[i];
+    for (uint32_t i = 0; i < n; i++) offset[i + 1] = offset[i] + deg[i];
+    uint32_t total_edges = offset[n];
+    uint32_t *adj = (uint32_t *)malloc(total_edges * sizeof(uint32_t));
+    uint32_t *pos = (uint32_t *)calloc(n, sizeof(uint32_t));
+    for (uint32_t i = 0; i < m; i++) {
+        uint32_t u = eu[i], v = ev[i];
         adj[offset[u] + pos[u]++] = v;
         adj[offset[v] + pos[v]++] = u;
     }
@@ -57,20 +57,20 @@ int main(int argc, char **argv) {
     auto t1 = Clock::now();
 
     // BFS from vertex 0
-    int *visited = (int *)calloc(n, sizeof(int));
-    int *queue = (int *)malloc(n * sizeof(int));
-    int qhead = 0, qtail = 0;
+    uint32_t *visited = (uint32_t *)calloc(n, sizeof(uint32_t));
+    uint32_t *queue = (uint32_t *)malloc(n * sizeof(uint32_t));
+    uint32_t qhead = 0, qtail = 0;
 
     visited[0] = 1;
     queue[qtail++] = 0;
 
     int64_t dist_sum = 0;
-    int *dist = (int *)calloc(n, sizeof(int));
+    int64_t *dist = (int64_t *)calloc(n, sizeof(int64_t));
 
     while (qhead < qtail) {
-        int v = queue[qhead++];
-        for (int i = offset[v]; i < offset[v + 1]; i++) {
-            int w = adj[i];
+        uint32_t v = queue[qhead++];
+        for (uint32_t i = offset[v]; i < offset[v + 1]; i++) {
+            uint32_t w = adj[i];
             if (!visited[w]) {
                 visited[w] = 1;
                 dist[w] = dist[v] + 1;
@@ -83,7 +83,7 @@ int main(int argc, char **argv) {
     double compute_ms = ms_since(t1);
 
     // ── Output ───────────────────────────────────────────────────────────────
-    printf("read=%.1fms compute=%.1fms checksum=%lld visited=%d\n",
+    printf("read=%.1fms compute=%.1fms checksum=%lld visited=%u\n",
            read_ms, compute_ms, (long long)dist_sum, qtail);
 
     free(visited); free(queue); free(dist); free(adj); free(offset);
